+++ callingcardstools/PeakCalling/yeast/poisson_pval.py
@@ -1,4 +1,5 @@
 import logging
+import warnings
 
 from scipy.stats import poisson
 
@@ -27,6 +28,8 @@ def poisson_pval(total_background_hops: int,
     :return: The Poisson p-value.
     :rtype: float
     """
+    warnings.warn("This function is deprecated and will be removed in a future release. "
+                  "use the vectorized function instead")
     # check input
     if total_background_hops < 0 or not isinstance(total_background_hops, int):
         raise ValueError(('total_background_hops must '